metrics-exporters/requirements.txt
prometheus-client
pyyaml
argparse
pyjson